feat: fix request for hero details
