+++ src/index.css
@@ -45,5 +45,5 @@ body {
   color: #252b42;
   font-family: Georgia, sans-serif;
   font-weight: 500;
-  background: rgb(113 118 145 / 60%);
+  background: rgb(179 181 193 / 60%);
 }
