+++ src/view/Hero/Hero.scss
@@ -5,7 +5,6 @@
   padding: 10px;
   border: 1px solid black;
   background-color: rgb(227 205 227);
-  margin-bottom: 10px;
   cursor: pointer;
 
   .hero__title {
