+++ src/view/Search/Search.scss
@@ -63,9 +63,6 @@
   grid-template-columns: 70% 30%;
   padding: 10px;
 
-  .searchItems__heroes {
-  }
-
   .searchItems__details {
   }
 }
@@ -74,4 +71,10 @@
   display: flex;
   justify-content: center;
   padding: 10px;
+}
+
+.searchItems__heroes {
+  display: flex;
+  flex-direction: column;
+  row-gap: 10px;
 }
